fix(app): deprecate the 'ts' extVar (#244)
The 'ts' extVar doesn't have much value and requires invocation of things like `date`. It's non-standard at Outreach. I propose we deprecate it and eventually remove the field.

Using the static value '1690825986' as the output of `date +%s` (one of the ways of generating it) of the deprecation time.

Updated the CODEOWNERS to have DT own `kubernetes/app.libsonnet`