fix(PodDisruptionBudget): ease setting all options
After restenciling, a new linter appears to have been enabled which
barfed on `{ maxUnavailable: null, minAvailable: 1 }`, because the
relevant JSON schema specifies `oneOf` for those two values.

Prior to this commit, PodDisruptionBudget's implementation makes it
difficult to satisfy that `oneOf` linter requirement, because removing
 keys from a JSonnet object is less than straightforward

This commit updates the function to accept a `specPartial` argument that
is added to the spec. Doing it this way ensures that

1. Clients using the defaults can continue to do so
2. Clients overriding the defaults can continue to do so
3. Clients which rely on `minAvailable` can set it without needing to
   write an iterator to filter out the old values. I think expecting
   clients to do that is likely to result in more bugs than simply
   updating this function here